requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic==2.10.6
sqlalchemy==2.0.27
openai>=1.66.2
tiktoken==0.6.0
python-multipart==0.0.9
openai-whisper==20231117
numpy==1.26.4
ffmpeg-python==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
aiosqlite==0.19.0
pytest==8.0.2
httpx==0.27.0
openai-agents==0.0.4